Factor out RefActor, Factor out datacollector, Factor out param server, Add trainer and logger
