+++ torchtune/dev/rl/datatypes/__init__.py
@@ -1,0 +1,3 @@
+from .request_output import RequestOutput
+from .trajectory import Trajectory
+from .vllm_completion_output import VllmCompletionOutput
