+++ torchtune/dev/rl/datatypes/vllm_completion_output.py
@@ -1,0 +1,4 @@
+import vllm
+from tensordict import from_dataclass
+
+VllmCompletionOutput = from_dataclass(vllm.outputs.CompletionOutput)
